requirements.txt
pandas
python-dotenv
openai
streamlit
Flask
